requirements.txt
fastmcp>=0.1.0
atlassian-python-api>=3.41.0
pydantic>=2.0.0
python-dotenv>=1.0.0
fuzzywuzzy>=0.18.0
python-levenshtein>=0.21.0
